Progress on tests
